requirements.txt
allure_python_commons==2.12.0
pytest==7.2.1
PyYAML==6.0
Requests==2.31.0
xlrd==2.0.1

